% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_movies.R
\name{count_movies}
\alias{count_movies}
\title{Count Movies by Field (by Actor, by Genre, by Year, etc.)}
\usage{
count_movies(count_by = NULL, path = ".")
}
\arguments{
\item{count_by}{The field to count movies by.}

\item{path}{The folder containing YAML files (or the root project).}
}
\value{
A 2-columns data frame.
}
\description{
This function counts movies by category: genre, year, actors, director,
writer, language or country.
}
\examples{
\dontrun{
count_movies()
count_movies(count_by = "genre")
count_movies(count_by = "actors")
}
}
