% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_imdb_id.R
\name{find_imdb_id}
\alias{find_imdb_id}
\title{Retrieve the IMDb Identifier of a Movie}
\usage{
find_imdb_id(search, year = NULL, n_max = 10, sleep = 0.1, verbose = TRUE)
}
\arguments{
\item{search}{A movie title to search for.}

\item{year}{(optional) The release year of the movie.}

\item{n_max}{Number of matches to returned (default is 10).}

\item{sleep}{Time interval between two API requests.}

\item{verbose}{A boolean. If TRUE, prints some informations.}
}
\value{
A data frame with:
\itemize{
\item title: the movie title
\item year: the year of release
\item imdbid: the IMDb ID of the movie
}
}
\description{
This function tries to retrieve the IMDb identifier of a movie based on a its
title and using the OMDb API \url{http://www.omdbapi.com/}. User can also
specify the release year of the movie. See details below for further
informations.
}
\details{
For better results, you should write the complete movie title with its exact
spelling. Results are sorted by their similarity with the search terms.

If you don't find the movie you are looking for, you can visit the IMDb
website (\url{https://www.imdb.com}) and get this ID from the movie URL. It
is always in the form: tt9999999 (only numbers are specific to the movie; the
prefix 'tt' is a constant).

For instance, in this URL: https://www.imdb.com/title/tt5699154/, the IMDb ID
is 'tt5699154'.

For non-english movies, the english name might be different from the original
title. For example, the english title of the french movie "Le Sens de la
fête" (2017) is "C'est la vie!" (IMDb ID = tt5699154).

Only movies are currently implemented.
}
\examples{
\dontrun{
find_imdb_id(search = "solo a star wars story")

find_imdb_id("Star Wars", year = 2018, n_max = 20)
}
}
