% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cover.R
\name{get_cover}
\alias{get_cover}
\title{Retrieve Cover/Poster of a Movie}
\usage{
get_cover(imdb_id, path = ".")
}
\arguments{
\item{imdb_id}{The IMDb ID of the movie.}

\item{path}{The folder to save cover.}
}
\description{
This function retrieves movie cover using the YTS API
\url{https://yts.mx/api}. No API key is required but not all movies listed in
the IMDb/OMDb databases are available on YTS \url{https://yts.mx}.
}
\examples{
\dontrun{
get_cover(imdb_id = "tt0120863", path = ".")
}
}
