% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Retrieves Metadata of a Movie}
\usage{
get_details(imdb_id, path = ".", print = TRUE)
}
\arguments{
\item{imdb_id}{The IMDb ID of the movie.}

\item{path}{The folder to store results.}

\item{print}{A boolean. If TRUE (default), movie informations are printed.}
}
\value{
A data frame with:
\itemize{
\item imdbid: the IMDb ID of the movie
\item type: the category (e.g. movie)
\item title: the movie title
\item year: the year of release
\item runtime: the movie runtime (in minutes)
\item director: a vector of directors
\item writer: a vector of writers
\item actors: a vector of main actors
\item genre: a vector of genres
\item plot: a short plot of the movie
\item language: a vector of spoken languages
\item country: a vector of the countries
\item imdbrating: the IMDb rating (in date of the request)
\item slug: a unique identifier (titre + year)
}
}
\description{
This function retrieves movie metadata (title, actors, year, genres, etc.)
using the OMDb API \url{http://www.omdbapi.com/}. Results are exported in a
YAML file and returned as a data frame. See details below for further
informations.
}
\details{
An (free) API key is required to use the OMDb API. You can register on
\url{http://www.omdbapi.com/}. When using this package for the first time,
you'll be asked for setting your own API key (just follow instructions).

The request is performed using the IMDb identifier of the movie. You can
visit the IMDb website (\url{https://www.imdb.com}) and get this ID from the
movie URL. It is always in the form: tt9999999 (only numbers are specific to
the movie; the prefix 'tt' is a constant).

For instance, in this URL: https://www.imdb.com/title/tt5699154/, the IMDb ID
is 'tt5699154'.

Another way to find the IMDb identifierof a movie is to use the function
\code{find_imdb_id()}, but it could must faster to directly visit the IMDb
website.

For non-english movies, the english name might be different from the original
title. For instance, the english title of the french movie "Le Sens de la
fête" (2017) is "C'est la vie!" (IMDb ID = tt5699154).

Only movies are currently implemented.
}
\examples{
\dontrun{
x <- get_details("tt0120863", path = ".", print = TRUE)
}
}
