% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_details.R
\name{read_details}
\alias{read_details}
\title{Read Metadata of Movie(s)}
\usage{
read_details(path = ".", imdb_id = NULL, print = TRUE)
}
\arguments{
\item{path}{The folder containing YAML files (or the root project).}

\item{imdb_id}{(optional) The IMDb ID(s) of the movie(s).}

\item{print}{A boolean. If TRUE (default), movie informations are printed (if
one single IMDb ID is required).}
}
\value{
A data frame with:
\itemize{
\item imdbid: the IMDb ID of the movie
\item type: the category (e.g. movie)
\item title: the movie title
\item year: the year of release
\item runtime: the movie runtime (in minutes)
\item director: a vector of directors
\item writer: a vector of writers
\item actors: a vector of main actors
\item genre: a vector of genres
\item plot: a short plot of the movie
\item language: a vector of spoken languages
\item country: a vector of the countries
\item imdbrating: the IMDb rating (in date of the request)
\item slug: a unique identifier (titre + year)
}
}
\description{
This function reads YAML files with movie(s) metadata stored by the function
\code{get_details()}
}
\examples{
\dontrun{
x <- read_details()
x <- read_details(imdb_id = "tt2278388", print = TRUE)
x <- read_details(imdb_id = c("tt2278388", "tt0362270"))
}
}
