% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_movies.R
\name{search_movies}
\alias{search_movies}
\title{Search for Movies}
\usage{
search_movies(
  title,
  year,
  genre,
  director,
  writer,
  actors,
  country,
  language,
  path = "."
)
}
\arguments{
\item{title}{Terms to search for in title field.}

\item{year}{Year(s) to search for in year field.}

\item{genre}{Terms to search for in genres field.}

\item{director}{Name(s) to search for in director field.}

\item{writer}{Name(s) to search for in writer field.}

\item{actors}{Name(s) to search for in actors field.}

\item{country}{Country names to search for in country field.}

\item{language}{Language name(s) to search for in language field.}

\item{path}{The folder containing YAML files (or the root project).}
}
\value{
A data frame.
}
\description{
This function searches for movies for which details have been stored. User
can search in title, year, etc.
}
\examples{
\dontrun{
search_movies(title = "Hotel")
search_movies(year = 2014:2016)
search_movies(actors = "Bill Murray")
}
}
