% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omdbr-package.R
\docType{package}
\name{omdbr-package}
\alias{omdbr}
\alias{omdbr-package}
\title{omdbr: Retrieve Movies Metadata from OMDb Database}
\description{
This package retrieves movies metadata (title, actors, year,
    genres, etc.) using the OMDb API <http://www.omdbapi.com/> (an API key is
    required). Cover can also be downloaded from the YTS API
    <https://yts.mx/api> (no API key required).
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com}

}
\keyword{internal}
